# docs/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "CyberAgent Workbench Local API",
    "description": "Authenticated loopback-only API owned by the Go control plane. Read operations expose durable metadata; the separately authorized control capability only requests exact active-call cancellation.",
    "version": "api.v1",
    "license": {
      "name": "Apache License 2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8765",
      "description": "Default loopback server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/notes/{note_id}": {
      "get": {
        "operationId": "getNote",
        "summary": "Inspect a Note",
        "description": "Returns one redacted structured Note.",
        "tags": [
          "Memory"
        ],
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "description": "Note identity",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[^/\\\\\\x00-\\x1f\\x7f]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful read",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/NoteView"
                    },
                    "request_id": {
                      "maxLength": 256,
                      "minLength": 1,
                      "type": "string"
                    },
                    "version": {
                      "const": "api.v1",
                      "type": "string"
                    }
                  },
                  "required": [
                    "version",
                    "request_id",
                    "data"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/ResourceExhausted"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "x-cyberagent-read-only": true
      }
    },
    "/api/v1/openapi.json": {
      "get": {
        "operationId": "getOpenAPI",
        "summary": "Read the OpenAPI contract",
        "description": "Returns the raw deterministic OpenAPI 3.1 JSON document under the same authentication boundary.",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "OpenAPI 3.1 document",
            "content": {
              "application/vnd.oai.openapi+json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/ResourceExhausted"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "x-cyberagent-read-only": true
      }
    },
    "/api/v1/runs": {
      "get": {
        "operationId": "listRuns",
        "summary": "List Runs",
        "description": "Returns a bounded cursor page of durable Runs.",
        "tags": [
          "Runs"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size from 1 to 100; defaults to 50",
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor bound to this route and exact filter set",
            "schema": {
              "maxLength": 512,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Exact Run status filter",
            "schema": {
              "enum": [
                "created",
                "preparing",
                "running",
                "waiting_approval",
                "paused",
                "completed",
                "failed",
                "cancelled"
              ],
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "mission_id",
            "in": "query",
            "description": "Exact Mission identity filter",
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[^/\\\\\\x00-\\x1f\\x7f]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful read",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/RunView"
                      },
                      "type": "array"
                    },
                    "page": {
                      "$ref": "#/components/schemas/Page"
                    },
                    "request_id": {
                      "maxLength": 256,
                      "minLength": 1,
                      "type": "string"
                    },
                    "version": {
                      "const": "api.v1",
                      "type": "string"
                    }
                  },
                  "required": [
                    "version",
                    "request_id",
                    "data",
                    "page"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/ResourceExhausted"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "x-cyberagent-read-only": true
      }
    },
...
    "/api/v1/runs/{run_id}/active-call/cancel": {
      "post": {
        "operationId": "requestModelCancellation",
        "summary": "Cancel an active model call",
        "description": "Persists an audit-first cancellation request bound to the exact active Supervisor and model attempt. The worker consumes it with its private execution lease; clients never provide a fencing token.",
        "tags": [
          "Control"
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "description": "Run identity",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[^/\\\\\\x00-\\x1f\\x7f]+$",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Opaque operation key; only a domain-separated digest is persisted",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 16,
              "pattern": "^\\S+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Cancellation request accepted or idempotently replayed",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ModelCancellationView"
                    },
                    "request_id": {
                      "maxLength": 256,
                      "minLength": 1,
                      "type": "string"
                    },
                    "version": {
                      "const": "api.v1",
                      "type": "string"
                    }
                  },
                  "required": [
                    "version",
                    "request_id",
                    "data"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/FailedPrecondition"
          },
          "413": {
            "$ref": "#/components/responses/RequestEntityTooLarge"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "429": {
            "$ref": "#/components/responses/ResourceExhausted"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "ControlBearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCancellationRequestView"
              }
            }
          }
        },
        "x-cyberagent-read-only": false
      }
    },
...
    "/api/v1/sessions/{session_id}": {
      "get": {
        "operationId": "getSession",
        "summary": "Inspect a Session",
        "description": "Returns Session metadata and its optional bound Run.",
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identity",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[^/\\\\\\x00-\\x1f\\x7f]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful read",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/SessionDetailView"
                    },
                    "request_id": {
                      "maxLength": 256,
                      "minLength": 1,
                      "type": "string"
                    },
                    "version": {
                      "const": "api.v1",
                      "type": "string"
                    }
                  },
                  "required": [
                    "version",
                    "request_id",
                    "data"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/ResourceExhausted"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "x-cyberagent-read-only": true
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NoteView": {
        "additionalProperties": false,
        "properties": {
          "archived_at": {
            "format": "date-time",
            "type": "string"
          },
          "category": {
            "enum": [
              "observation",
              "hypothesis",
              "decision",
              "summary",
              "reference"
            ],
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "maxLength": 256,
            "minLength": 1,
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "owner_agent_id": {
            "maxLength": 256,
            "minLength": 1,
            "type": "string"
          },
          "pinned": {
            "type": "boolean"
          },
          "run_id": {
            "maxLength": 256,
            "minLength": 1,
            "type": "string"
          },
          "source_refs": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "status": {
            "enum": [
              "active",
              "archived"
            ],
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          },
          "visibility": {
            "enum": [
              "run",
              "root",
              "owner"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "title",
          "content",
          "category",
          "visibility",
          "tags",
          "source_refs",
          "evidence_ids",
          "status",
          "pinned",
          "version",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Process-scoped read token; never persisted by CyberAgent."
      },
      "ControlBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Distinct optional control token; cannot authorize read operations and is never persisted."
      }
    }
  },
...
}